Add rewards module
