+++ programs/stake/src/lib.rs
@@ -13,6 +13,7 @@ use solana_sdk::{
 
 pub mod config;
 pub mod points;
+pub mod rewards;
 pub mod stake_instruction;
 pub mod stake_state;
